LIBITD-1286. Removed unnecessary configuration from test_helper.rb
Removed old pre-Rails 5.1 configuration from test/test_helper.rb
method.

Moved some of the browser-focused methods from
ActionDispatch::IntegrationTest into the ApplicationSystemTestCase
class.

https://issues.umd.edu/browse/LIBITD-1286
